requirements.txt
pyarrow<10.1.0,>=10.0.1
pandas
numpy
boto3
requests
apache-airflow-providers-ssh
apache-airflow-providers-amazon
mlflow
numpy
cloudpickle==2.0.0
scikit-learn==1.0.2
evidently
pymongo